Moved SkillIconPanel into its own class
